Reorganize helpers
Having helpers grouped by category does not aid findability; on the contrary, it makes finding and linking helpers more difficult.
