+++ source/components/checkbox.scss
@@ -2,7 +2,7 @@
 // -----------------------------------------------------------------------------
 
 @use "../settings" as *;
-@use "../helpers/accessibility/focus-visible" as *;
+@use "../helpers/focus-visible" as *;
 
 .checkbox {
   width: 1.5rem;
