+++ source/components/popover.js
@@ -1,7 +1,7 @@
 // POPOVER
 // -----------------------------------------------------------------------------
 
-import togglePoppable from "../helpers/miscellaneous/toggle-poppable";
+import togglePoppable from "../helpers/toggle-poppable";
 
 const SELECTOR_POPOVER = ".popover";
 const CLASS_ACTIVATED = "is-activated";
