+++ source/helpers/animate-fade.scss
@@ -1,4 +1,4 @@
-// FADE
+// ANIMATE FADE
 // -----------------------------------------------------------------------------
 
 @keyframes fade-in-regular {
