+++ source/helpers/animate-slide.scss
@@ -1,7 +1,7 @@
-// SLIDE
+// ANIMATE SLIDE
 // -----------------------------------------------------------------------------
 
-@use "../../settings" as *;
+@use "../settings" as *;
 
 @keyframes slide-in-regular {
   from {
