+++ source/helpers/trap-focus.js
