+++ source/parts/local-navigation.scss
@@ -2,7 +2,7 @@
 // -----------------------------------------------------------------------------
 
 @use "../settings" as *;
-@use "../helpers/accessibility/focus-visible" as *;
+@use "../helpers/focus-visible" as *;
 
 .local-navigation {
   display: flex;
